refactor: combine rsk3 and web3 tests
